# server/docs/swagger.yaml
basePath: /
definitions:
...
  types.AdminUserDetail:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_global_admin:
        type: boolean
      linked:
        type: boolean
      memberships:
        items:
          $ref: '#/definitions/types.UserMembershipView'
        type: array
      verification_status:
        $ref: '#/definitions/types.VerificationStatus'
    type: object
...
  types.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  types.LogoutRequest:
    properties:
      refresh_token:
        type: string
    type: object
...
  types.MembershipRole:
    enum:
    - group_admin
    - member
    type: string
    x-enum-varnames:
    - RoleGroupAdmin
    - RoleMember
  types.MembershipStatus:
    enum:
    - requested
    - approved
    - rejected
    type: string
    x-enum-varnames:
    - MembershipRequested
    - MembershipApproved
    - MembershipRejected
...
  types.Page:
    properties:
      items: {}
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  types.Principal:
    properties:
      email:
        type: string
      id:
        type: string
      is_global_admin:
        type: boolean
      verification_status:
        $ref: '#/definitions/types.VerificationStatus'
    type: object
...
  types.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  types.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
  types.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/types.Principal'
    type: object
...
  types.UserMembershipView:
    properties:
      created_at:
        type: string
      group_id:
        type: string
      group_name:
        type: string
      role:
        $ref: '#/definitions/types.MembershipRole'
      status:
        $ref: '#/definitions/types.MembershipStatus'
    type: object
  types.VerificationStatus:
    enum:
    - registered
    - pending_verification
    - verified
    - rejected
    type: string
    x-enum-varnames:
    - VerificationRegistered
    - VerificationPending
    - VerificationVerified
    - VerificationRejected
  types.apiError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    Shared-expense settlement tracker for group trips. Members record what they paid; closing a group computes each member's fair share and the minimum set of payments that squares everyone up. Money moves off-platform — each computed payment is tracked through a two-key confirmation flow (debtor proves, creditor attests, admin finalizes).

    All amounts are integer **baisa** (1.000 OMR = 1000). There are no floats in the money path.

    **Authenticating here:** call `POST /auth/login` (try `admin@expense-splitter.local` / `admin`), copy the `access_token` from the response, click **Authorize** above and paste `Bearer <access_token>`.
  title: Expense Splitter API
  version: "1.0"
paths:
  /admin/groups:
    get:
      description: Global admin only. Every group in the system, regardless of membership.
        Returns a `{total, limit, offset, items}` envelope.
      parameters:
      - default: 50
        description: page size (1-200)
        in: query
        name: limit
        type: integer
      - default: 0
        description: rows to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: paginated groups
          schema:
            $ref: '#/definitions/types.Page'
        "400":
          description: invalid paging
          schema:
            $ref: '#/definitions/types.apiError'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/types.apiError'
        "403":
          description: global admin only
          schema:
            $ref: '#/definitions/types.apiError'
      security:
      - BearerAuth: []
      summary: List all groups
      tags:
      - admin
  /admin/groups/{id}:
    delete:
      description: Global admin only, and **pristine groups only** — a group with
        any recorded expense or computed payment cannot be deleted.
      parameters:
      - description: group id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: group deleted
        "400":
          description: invalid group id
          schema:
            $ref: '#/definitions/types.apiError'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/types.apiError'
        "403":
          description: global admin only
          schema:
            $ref: '#/definitions/types.apiError'
        "404":
          description: group not found
          schema:
            $ref: '#/definitions/types.apiError'
        "409":
          description: group is not pristine
          schema:
            $ref: '#/definitions/types.apiError'
      security:
      - BearerAuth: []
      summary: Delete a group
      tags:
      - admin
  /admin/users:
    get:
      description: Global admin only. Filter by verification status to work the approval
        queue. Returns a `{total, limit, offset, items}` envelope.
      parameters:
      - description: filter by verification status
        enum:
        - registered
        - pending_verification
        - verified
        - rejected
        in: query
        name: status
        type: string
      - default: 50
        description: page size (1-200)
        in: query
        name: limit
        type: integer
      - default: 0
        description: rows to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: paginated users
          schema:
            $ref: '#/definitions/types.Page'
        "400":
          description: invalid status or paging
          schema:
            $ref: '#/definitions/types.apiError'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/types.apiError'
        "403":
          description: global admin only
          schema:
            $ref: '#/definitions/types.apiError'
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - admin
  /admin/users/{id}:
    delete:
      description: Global admin only. Refused when the user has recorded expenses
        or is party to a payment — removing them would corrupt a split.
      parameters:
      - description: user id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: user deleted
        "400":
          description: invalid user id
          schema:
            $ref: '#/definitions/types.apiError'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/types.apiError'
        "403":
          description: global admin only
          schema:
            $ref: '#/definitions/types.apiError'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/types.apiError'
        "409":
          description: user has expenses or payments
          schema:
            $ref: '#/definitions/types.apiError'
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - admin
    get:
      description: Global admin only. The account plus its group memberships and roles.
      parameters:
      - description: user id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: the user and their memberships
          schema:
            $ref: '#/definitions/types.AdminUserDetail'
        "400":
          description: invalid user id
          schema:
            $ref: '#/definitions/types.apiError'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/types.apiError'
        "403":
          description: global admin only
          schema:
            $ref: '#/definitions/types.apiError'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/types.apiError'
      security:
      - BearerAuth: []
      summary: Get one user's full detail
      tags:
      - admin
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Exchanges email + password for an access token (~5 min) and a refresh
        token. Also returns your local `user` row, which is auto-provisioned/linked
        on first login. Copy the `access_token` into the **Authorize** button to use
        the rest of this page.
      parameters:
      - description: email + password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: access_token, refresh_token, user
          schema:
            $ref: '#/definitions/types.TokenResponse'
        "400":
          description: invalid body
          schema:
            $ref: '#/definitions/types.apiError'
        "401":
          description: bad credentials
          schema:
            $ref: '#/definitions/types.apiError'
      summary: Log in
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revokes the session behind the given refresh token. Idempotent
        — logging out an already-revoked session still succeeds.
      parameters:
      - description: refresh_token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.LogoutRequest'
      produces:
      - application/json
      responses:
        "204":
          description: session revoked
        "400":
          description: invalid body
          schema:
            $ref: '#/definitions/types.apiError'
      summary: Log out
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Renews the access/refresh token pair without re-sending the password.
        Fails once the session has been revoked by logout.
      parameters:
      - description: refresh_token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: a fresh token pair
          schema:
            $ref: '#/definitions/types.TokenResponse'
        "400":
          description: invalid body
          schema:
            $ref: '#/definitions/types.apiError'
        "401":
          description: refresh token is invalid or expired
          schema:
            $ref: '#/definitions/types.apiError'
      summary: Refresh a session
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates the Keycloak identity and the linked local user row. The
        new account starts at verification_status `registered` — it must be verified
        before it can join groups or record expenses.
      parameters:
      - description: email, password, name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: the created account
          schema:
            $ref: '#/definitions/types.Principal'
        "400":
          description: invalid body / weak password
          schema:
            $ref: '#/definitions/types.apiError'
        "409":
          description: email already registered
          schema:
            $ref: '#/definitions/types.apiError'
      summary: Sign up a new user
      tags:
      - auth
...
  /groups/{id}/audit:
    get:
      description: 'Admins only. The full history: expense edits with **before/after**
        amounts, role transfers, membership decisions, and every payment-state transition.
        Returns a `{total, limit, offset, items}` envelope.'
      parameters:
      - description: group id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - default: 50
        description: page size (1-200)
        in: query
        name: limit
        type: integer
      - default: 0
        description: rows to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: paginated audit entries
          schema:
            $ref: '#/definitions/types.Page'
        "400":
          description: invalid id or paging
          schema:
            $ref: '#/definitions/types.apiError'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/types.apiError'
        "403":
          description: not an admin of this group
          schema:
            $ref: '#/definitions/types.apiError'
        "404":
          description: group not found
          schema:
            $ref: '#/definitions/types.apiError'
      security:
      - BearerAuth: []
      summary: Read a group's audit trail
      tags:
      - ops
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Bearer token from POST /auth/login. Format: `Bearer <access_token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...